Finish TablePress 2.1.5.

+++ views/class-editor-button-thickbox-list-table.php
@@ -337,7 +337,7 @@ class TablePress_Editor_Button_Thickbox_List_Table extends WP_List_Table {
 		$this->set_pagination_args( array(
 			'total_items' => $total_items, // Total number of records/items.
 			'per_page'    => $per_page, // Number of items per page.
-			'total_pages' => ceil( $total_items / $per_page ), // Total number of pages.
+			'total_pages' => (int) ceil( $total_items / $per_page ), // Total number of pages.
 		) );
 	}
